[JVM] Rewrite `isCallInsideSameModuleAsCallee`
Use `IrModuleFragment` instead of `ModuleDescriptor`. This change is
required for the evaluate expression in IDE. When we compile
some code fragment with inline function call that has an anonymous
object in callee, we will get incorrect behavior. Code fragment is
wrapped in `EvaluatorModuleDescriptor` and we accidentally
think that inline call and callee are in different modules that
leads to an error in `AnonymousObjectTransformer.doTransform`.