fixup! [K1] Fix false negative DEPRECATED_RESOLVE_WITH_AMBIGUOUS_ENUM_ENTRY
Change the API. It looks like trying to
find a single classifier is OK in some
cases: for them there are the
`getContributedClassifier(?!s)*`
functions.

In some cases it seems that rewriting
the code to account for multiple
classifiers may be the correct thing:
in these cases an explicit
`.singleOrNull()` is used, so that they
(1) stand out in the commit/review
(2) stand out in the code by looking
weird.

Note that this commit should preserve
the semantics: we shouldn't "submit"
more classifiers at any given scope
compared to how many we used to,
but we are allowed to process more
classifiers instead of just one. This is
true, because we de-facto shouldn't
be able to get more than what we used to
have.